数据标签module：数据标签计算sql组装

+++ idata-label/src/main/java/cn/zhengcaiyun/idata/label/service/LabObjectLabelService.java
@@ -51,5 +51,5 @@ public interface LabObjectLabelService {
      * @param layerId
      * @return
      */
-    LabelQueryDataDto queryLabelResultData(Long id, Long layerId, Integer limit, Integer offset);
+    LabelQueryDataDto queryLabelResultData(Long id, Long layerId, Long limit, Long offset);
 }
